+++ beavermoney.graphql
@@ -62,6 +62,7 @@ input TimePeriodInput {
   startDate: Time
   endDate: Time
   preset: TimePeriodPreset
+  timezone: String!
 }
 
 enum TimePeriodPreset {
